fix(providers): align lazy adapter contracts
